+++ NEWS.md
@@ -3,6 +3,8 @@
 * Implemented simple deduplication of files with the same name and size across
   OS images. (#19)
 * Deduplication is triggered automatically at the end of `monci update` (#19)
+* Allow to configure a compression property to use when creating BTRFS
+  subvolumes
 
 # Version 0.1